feat(Prerender): Set a timeout on prerender ready.
While doing this I noticed how slow the apply page could be. Fractal
apply page took ~40s to load for me because of the project image. I
rearranged the page so that the project image loading wouldn't block and
now it loads in ~5s. I also hooked up the prerendering in a driver so
that the page could indicate that it is ready in less than the usual 30s
timeout.