fix(openAndPrint): use the current page for printing (#45)
The open and print driver will now use the current page, including
all of it's loaded CSS, to ensure that the part of the page we want
to render is styled correctly.
There is a small hack between the openAndPrint driver and the DOM driver
in order to ensure that event handlers continue to work. For instance, if
you do not 'rerender', the print button will not work after closing the print
dialogue until moving to another page first.