feat(csv): generate csvs
Allow feature of generating CSVs for admins only.
By navigating to /admin/csv you will be provided a list of projects,
identical to /admin/projects minus the project creation field. When an admin
clicks on one of the projects a csv will be generated and downloaded as sparks-admin.csv
which will increment based on your browsers download behavior. e.g sparks-admin(1).csv the
second time you download and sparks-admin(2).csv the third etc.